+++ tests/system/task.py
@@ -2,6 +2,7 @@
 
 import collections
 import datetime
+import unittest.mock
 
 import HABApp.openhab.items
 
@@ -11,8 +12,8 @@ import tests.helper.oh_item
 import tests.helper.test_case_base
 
 
-class TestRecurringTaskConfig(tests.helper.test_case_base.TestCaseBase):
-    """Tests for RecurringTaskConfig Rule."""
+class TestRecurringTask(tests.helper.test_case_base.TestCaseBase):
+    """Tests for RecurringTask Rule."""
 
     def setUp(self) -> None:
         """Setup test case."""
@@ -21,11 +22,11 @@ class TestRecurringTaskConfig(tests.helper.test_case_base.TestCaseBase):
         tests.helper.oh_item.add_mock_item(HABApp.openhab.items.SwitchItem, "Unittest_Task", None)
         tests.helper.oh_item.add_mock_item(HABApp.openhab.items.DatetimeItem, "Unittest_Task_last", None)
 
-        config = habapp_rules.system.config.task.RecurringTaskConfig(
+        config_max = habapp_rules.system.config.task.RecurringTaskConfig(
             items=habapp_rules.system.config.task.RecurringTaskItems(task_active="Unittest_Task", last_done="Unittest_Task_last"), parameter=habapp_rules.system.config.task.RecurringTaskParameter(recurrence_time=datetime.timedelta(hours=12))
         )
 
-        self._rule = habapp_rules.system.task.RecurringTask(config)
+        self._rule = habapp_rules.system.task.RecurringTask(config_max)
 
     def test_init_with_fixed_check_time(self) -> None:
         """Test init with fixed check time."""
@@ -37,6 +38,22 @@ class TestRecurringTaskConfig(tests.helper.test_case_base.TestCaseBase):
         )
 
         self.assertEqual(self._rule._config.parameter.fixed_check_time, datetime.time(7))
+
+    def test_init_with_min_config(self) -> None:
+        """Test init with minimal config."""
+        config_min = habapp_rules.system.config.task.RecurringTaskConfig(
+            items=habapp_rules.system.config.task.RecurringTaskItems(
+                task_active="Unittest_Task",
+            ),
+            parameter=habapp_rules.system.config.task.RecurringTaskParameter(recurrence_time=datetime.timedelta(hours=12)),
+        )
+
+        tests.helper.oh_item.add_mock_item(HABApp.openhab.items.DatetimeItem, "H_Unittest_Task_last_done", None)
+
+        with unittest.mock.patch("habapp_rules.core.helper.create_additional_item") as create_item_mock:
+            habapp_rules.system.task.RecurringTask(config_min)
+
+        create_item_mock.assert_called_once_with("H_Unittest_Task_last_done", "DateTime")
 
     def test__get_check_cycle(self) -> None:
         """Test _get_check_cycle()."""
@@ -81,3 +98,65 @@ class TestRecurringTaskConfig(tests.helper.test_case_base.TestCaseBase):
         tests.helper.oh_item.set_state("Unittest_Task_last", datetime.datetime.now() - datetime.timedelta(hours=1))
         self._rule._check_and_set_task_undone()
         tests.helper.oh_item.assert_value("Unittest_Task", "OFF")
+
+
+class TestCounterTask(tests.helper.test_case_base.TestCaseBase):
+    """Tests for CounterTask Rule."""
+
+    def setUp(self) -> None:
+        """Setup test case."""
+        tests.helper.test_case_base.TestCaseBase.setUp(self)
+
+        tests.helper.oh_item.add_mock_item(HABApp.openhab.items.SwitchItem, "Unittest_Counter_Task", None)
+        tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "Unittest_Observed", None)
+        tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "Unittest_Observed_last_reset", None)
+
+        config_max = habapp_rules.system.config.task.CounterTaskConfig(
+            items=habapp_rules.system.config.task.CounterTaskItems(task_active="Unittest_Counter_Task", observed="Unittest_Observed", last_reset="Unittest_Observed_last_reset"),
+            parameter=habapp_rules.system.config.task.CounterTaskParameter(max_value=42),
+        )
+
+        self._rule = habapp_rules.system.task.CounterTask(config_max)
+
+    def test_init_with_min_config(self) -> None:
+        """Test init with minimal config."""
+        config_min = habapp_rules.system.config.task.CounterTaskConfig(
+            items=habapp_rules.system.config.task.CounterTaskItems(task_active="Unittest_Counter_Task", observed="Unittest_Observed"), parameter=habapp_rules.system.config.task.CounterTaskParameter(max_value=42)
+        )
+
+        tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "H_Unittest_Observed_last_reset", None)
+
+        with unittest.mock.patch("habapp_rules.core.helper.create_additional_item") as create_item_mock:
+            habapp_rules.system.task.CounterTask(config_min)
+
+        create_item_mock.assert_called_once_with("H_Unittest_Observed_last_reset", "Number")
+
+    def test_overall_behaviour(self) -> None:
+        """Test overall behaviour."""
+        # first value
+        tests.helper.oh_item.item_state_change_event("Unittest_Observed", 20)
+        tests.helper.oh_item.assert_value("Unittest_Counter_Task", "OFF")
+
+        # second value still smaller than threshold
+        tests.helper.oh_item.item_state_change_event("Unittest_Observed", 42)
+        tests.helper.oh_item.assert_value("Unittest_Counter_Task", "OFF")
+
+        # value greater than threshold
+        tests.helper.oh_item.item_state_change_event("Unittest_Observed", 43)
+        tests.helper.oh_item.assert_value("Unittest_Counter_Task", "ON")
+
+        # second value greater than threshold
+        tests.helper.oh_item.item_state_change_event("Unittest_Observed", 100)
+        tests.helper.oh_item.assert_value("Unittest_Counter_Task", "ON")
+
+        # reset of task
+        tests.helper.oh_item.item_state_change_event("Unittest_Counter_Task", "OFF")
+        tests.helper.oh_item.assert_value("Unittest_Observed_last_reset", 100)
+
+        # first value after reset
+        tests.helper.oh_item.item_state_change_event("Unittest_Observed", 142)
+        tests.helper.oh_item.assert_value("Unittest_Counter_Task", "OFF")
+
+        # second value after reset
+        tests.helper.oh_item.item_state_change_event("Unittest_Observed", 143)
+        tests.helper.oh_item.assert_value("Unittest_Counter_Task", "ON")